[#1087] Moved FlashMessage component to its own directory.
