+++ frontend/src/common/flash-message/FlashMessage.less
